Created ProjectsList and Project component, refactored App component
